update python plotter
