+++ node_app/server/routes/generate.js
@@ -3,9 +3,9 @@ const express = require('express');
 const router = express.Router();
 const multer = require('multer');
 const path = require("path");
+const runPython = require("../controllers/run-python")
 
-// To run python script
-const {spawn} = require('child_process');
+
 
 const audioFile = {};
 
@@ -36,41 +36,20 @@ router.get('/spectrogram', async (req, res) => {
     if (audioFile.file) {
         res.render('generate/spectrogram', {filepath: audioFile.file.filename});
         // TODO: make link to results active after results are generated
-        // await new Promise(resolve => setTimeout(resolve, 5000));
+        // await new Promise(resolve => setTimeout(resolve, 1000));
     }
     else
         res.redirect(302, '/');
 });
 
-router.get('/results', (req, res) => {
+router.get('/results', async (req, res) => {
     if(audioFile.file) {
         fs.unlink(path.join(__dirname, '../../client/public/uploads/' + audioFile.file.filename), (err) => {
             if (err) throw err;
         });
     }
 
-    //
-    let python_response_code = 2;
-    const image_file = path.join(__dirname, '../../client/public/uploads/test.png');
-
-    // spawn new child process to call the python script
-    const python = spawn('python3', [path.join(__dirname,'../../../python_app/test_heroku.py'), image_file]);
-    // collect data from script
-    python.stdout.on('data', function (data) {
-        console.log(`Pipe data from python script ...${data}`);
-    });
-
-    python.stderr.on('data', function (data) {
-        console.log(`Pipe data from python script ...${data}`);
-    });
-
-    // in close event we are sure that stream from child process is closed
-    python.on('close', (code) => {
-        console.log(`child process close all stdio with code ${code}`);
-        python_response_code = code;
-        // send data to browser
-        //res.send(dataToSend)
-    });
+    await runPython.drawPythonPlot();
 
     // await new Promise(resolve => setTimeout(resolve, 3000));
